training container/requirements.txt
pandas
numpy
datetime
tensorflow
keras
json_numpy
google-cloud-storage
google-cloud-bigquery
db_dtypes